[TUA-833] [Home page] [user is not logged in] Verify that on the carousel slides are changing automatically
